# openapi.yml
openapi: '3.1.0'
info:
  title: Comprehensive Books Management REST API
  description: >-
    **A REST API for managing Users, Books, Authors, Genres, and their interactions, developed with [Express](https://expressjs.com).**

    This API provides endpoints to manage Users, Books, Authors, Genres, and the relationships between them.
  version: 1.0.0
  contact:
    url: https://github.com/Broannn/BookNest-RestAPI
servers:
  - url: https://booknest-restapi.onrender.com/
...
paths:
...
  /authors:
    post:
      summary: Create an author.
      tags:
        - authors
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorData'
      responses:
        201:
          description: The author was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
    get:
      summary: List all authors.
      tags:
        - authors
      responses:
        200:
          description: A list of authors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
  /authors/{authorId}:
    get:
      summary: Get an author by ID.
      tags:
        - authors
      parameters:
        - name: authorId
          in: path
          required: true
          description: The ID of the author.
          schema:
            type: string
      responses:
        200:
          description: The author details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        404:
          description: Author not found.
    put:
      summary: Update an author by ID.
      tags:
        - authors
      parameters:
        - name: authorId
          in: path
          required: true
          description: The ID of the author to update.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorData'
      responses:
        200:
          description: Author successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        404:
          description: Author not found.
    delete:
      summary: Delete an author by ID.
      tags:
        - authors
      parameters:
        - name: authorId
          in: path
          required: true
          description: The ID of the author to delete.
          schema:
            type: string
      responses:
        204:
          description: Author successfully deleted.
        404:
          description: Author not found.
components:
...
  schemas:
...
    Author:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the author.
          example: 12345
        name:
          type: string
          description: The name of the author.
          example: John Doe
        bio:
          type: string
          description: A brief biography of the author.
          example: "An accomplished author."
        birthDate:
          type: string
          format: date
          description: The birth date of the author.
          example: 1980-01-01
        createdAt:
          type: string
          format: date-time
          description: The date when the author was created.
          example: 2025-01-13T10:00:00.000Z
      required:
        - id
        - name
    AuthorData:
      type: object
      properties:
        name:
          type: string
          description: The name of the author.
          example: John Doe
        bio:
          type: string
          description: A brief biography of the author.
          example: "An accomplished author."
        birthDate:
          type: string
          format: date
          description: The birth date of the author.
          example: 1980-01-01
      required:
        - name
...